Moved GameMigrator "stuff" into Games directory
