+++ Distroir.CustomSDKLauncher.Core/Migrators/Games/GameCache.cs
@@ -8,7 +8,7 @@ using System.Text;
 using System.Windows.Forms;
 using System.Xml;
 
-namespace Distroir.CustomSDKLauncher.Core.Migrators
+namespace Distroir.CustomSDKLauncher.Core.Migrators.Games
 {
     public class GameCache
     {
