+++ Distroir.CustomSDKLauncher.Core/Migrators/Games/GameMigrationConflictSolution.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-namespace Distroir.CustomSDKLauncher.Core.Migrators
+namespace Distroir.CustomSDKLauncher.Core.Migrators.Games
 {
     public enum GameMigrationConflictSolution
     {
